Removed the all contracts
