--- contracts/ckb-std-tests/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "ckb-std-tests"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ckb-std = { path = "../../", features = [ "dlopen-c", "dummy-atomic", "log"] }
-blake2b-ref = { version = "0.3", default-features = false }
-bytes = { version = "1.7", default-features = false }
-log = { version = "0.4", default-features = false }
