--- contracts/ckb-std-tests/src/code_hashes.rs
@@ -1,6 +1,0 @@
-// this is a blake2b hash of the shared library `test/shared-lib/shared-lib.so`
-// remember to update this hash if you change the shared library or upgrade the toolchain
-pub const CODE_HASH_SHARED_LIB: [u8; 32] = [
-    225, 15, 172, 171, 171, 161, 4, 34, 170, 83, 228, 192, 169, 205, 29, 189, 112, 156, 164, 73,
-    222, 132, 215, 23, 163, 71, 194, 190, 4, 227, 187, 120,
-];
