--- contracts/ckb-std-tests/src/main.rs
@@ -1,27 +1,0 @@
-//! Generated by capsule
-//!
-//! `main.rs` is used to define rust lang items and modules.
-//! See `entry.rs` for the `main` function.
-//! See `error.rs` for the `Error` type.
-
-#![no_std]
-#![no_main]
-
-// define modules
-mod code_hashes;
-mod entry;
-mod error;
-
-use ckb_std::default_alloc;
-
-ckb_std::entry!(program_entry);
-default_alloc!();
-
-/// program entry
-fn program_entry() -> i8 {
-    // Call main function and return error code
-    match entry::main() {
-        Ok(_) => 0,
-        Err(err) => err as i8,
-    }
-}
